==> notebooks/01_preprocess/01_preprocessing.ipynb <==
# Notebook for converting the screenings data from DX to a pandas DataFrame and performs some basic preprocessing to make the data more managable

# Allows imported code to be updated without the need of running the import statement after each update
%load_ext autoreload
%autoreload 2

import os
import numpy as np
import pandas as pd

from time import time
from glob import glob
from joblib import load, dump
from tmdbv3api import TMDb, Movie

# helper file in directory
from helper import rm_mf

# Folder for storing information generated throughout the project
data_folder = '../data/'

## Setting up the TMDb connection

with open('C:/Users/marku/OneDrive/Skrivebord/TMDBkeys.txt') as keys:
    application_key = keys.readline().strip()
tmdb = TMDb()
tmdb.api_key = application_key
tmdb.language = 'en'
movie = Movie()

## Get all unique stripped movie titles

Converts all titles to lowercase and removes the version (3D, atmos, HFR, etc.)

If the code has been run and the generated list of stripped titles was stored then you can go down to the cell where you load the file

# Folder with screenings from DX
dx_folder = 'K:/team/dxcn.master_uis/data/'

# English names for the columns
columns = pd.Index([
    'std_movienr',
    'full_title',
    'date',
    'showtime',
    'sold',
    'room',
    'seats',
    'seats_b',
    'coverage'
])

%%time
movies = set()
for file in os.listdir(dx_folder):
    # Reads the files from the DX folder
    showings = pd.read_json(dx_folder+file, encoding='utf-8')
    # Skip if the file is empty
    if showings.shape == (0, 0): continue
    # Rename all the columns (originally in Norwegian)
    showings.columns = columns
    for _, i in showings.iterrows():
        movies.add((i['date'][:4], rm_mf(i['full_title'].lower())))

### Saves the file with the stripped titles

dump(movies, data_folder+'stripped_movie_titles.joblib')

### Loads in the file with the stripped titles

No need to run the code each time

movies = load(data_folder+'stripped_movie_titles.joblib')

## Gather additional metadata for all movies in our dataset

def set_dict(r, diff):
    """
    Function used to minimize duplicate code in the following code cell.
    It sets information regarding a movie into a dictionary for later use.
    """
    details = movie.details(r.id)
    return {
        'diff':                 diff,
        'title':                r.title,
        'original_title':       r.original_title,
        'release_date':         r.release_date,
        'genre_ids':            r.genre_ids,
        'overview':             r.overview,
        'original_language':    r.original_language,
        'popularity':           r.popularity,
        'vote_count':           r.vote_count,
        'vote_average':         r.vote_average,
        'tmdb_id':              r.id,
        'runtime':              details.runtime,
        'imdb_id':              details.imdb_id,
        'overview':             details.overview,
        'budget':               details.budget,
        'production_companies': details.production_companies,
        'revenue':              details.revenue,
        'tagline':              details.tagline,
        'cast':                 details.casts['cast'],
        'crew':                 details.casts['crew'],
        'keywords':             details.keywords
    }

%%time
# This cell is in desperate need of refactoring

# We start by creating a list of dictionaries rather than directly using a pandas
# DataFrame because this is more efficient when sequentially appending entries
movie_details = list()
for i in movies:
    if i[1] == '': continue
    try:
        s = movie.search(i[1])
    except:
        continue
    if len(s) == 0:
        continue
    elif len(s) == 1:
        try:
            # movie.details retrieves additional information that movie.search does not
            details = movie.details(s[0].id)
            # Structures the retreived information into a dictionary
            movie_details.append({
                'title':                     i[1],
                'show_year':                 i[0],
                'tmdb_title':                s[0].title,
                'tmdb_original_title':       s[0].original_title,
                'tmdb_genre_ids':            s[0].genre_ids,
                'tmdb_overview':             s[0].overview,
                'tmdb_original_language':    s[0].original_language,
                'tmdb_popularity':           s[0].popularity,
                'tmdb_vote_count':           s[0].vote_count,
                'tmdb_vote_average':         s[0].vote_average,
                'tmdb_release_date':         s[0].release_date,
                'tmdb_id':                   s[0].id,
                'tmdb_runtime':              details.runtime,
                'imdb_id':                   details.imdb_id,
                'tmdb_overview':             details.overview,
                'tmdb_budget':               details.budget,
                'tmdb_production_companies': details.production_companies,
                'tmdb_revenue':              details.revenue,
                'tmdb_tagline':              details.tagline,
                'tmdb_cast':                 details.casts['cast'],
                'tmdb_crew':                 details.casts['crew'],
                'tmdb_keywords':             details.keywords
            })
        except:
            continue
    elif len(s) > 1:
        show_year = int(i[0])
        
        m = dict()
        for r in s:
            # If the original title retrieved from TMDb is the exact same as the 
            # stripped title from DX then we assume that it is the correct movie
            if r.original_title.lower() == i[1]:
                m = set_dict(r, diff)
                break
            try:
                release_year = int(r.release_date[:4])
            except:
                continue
            # diff is used to check the relative difference between the release
            # of the movies and the showtime of the screening, we assume that
            # the lower the difference the higher the possibility that the 
            # retreived movie is the correct one
            diff = show_year - release_year
            if diff < 0: continue
            if 'diff' not in m:
                m = set_dict(r, diff)
            elif diff < m['diff']:
                m = set_dict(r, diff)
        if 'title' not in m: continue
        movie_details.append({
            'title':                     i[1],
            'show_year':                 i[0],
            'tmdb_title':                m['title'],
            'tmdb_original_title':       m['original_title'],
            'tmdb_release_date':         m['release_date'],
            'tmdb_genre_ids':            m['genre_ids'],
            'tmdb_overview':             m['overview'],
            'tmdb_original_language':    m['original_language'],
            'tmdb_popularity':           m['popularity'],
            'tmdb_vote_count':           m['vote_count'],
            'tmdb_vote_average':         m['vote_average'],
            'tmdb_id':                   m['tmdb_id'],
            'tmdb_runtime':              m['runtime'],
            'imdb_id':                   m['imdb_id'],
            'tmdb_overview':             m['overview'],
            'tmdb_budget':               m['budget'],
            'tmdb_production_companies': m['production_companies'],
            'tmdb_revenue':              m['revenue'],
            'tmdb_tagline':              m['tagline'],
            'tmdb_cast':                 m['cast'],
            'tmdb_crew':                 m['crew'],
            'tmdb_keywords':             m['keywords']
        })

# Transform the list of dictionaries to a Pandas DataFrame
movie_det = pd.DataFrame(movie_details)

movie_det.head()

### Save the collected data to file

dump(movie_det, data_folder+'movie_details.joblib')

movie_det = load(data_folder+'movie_details.joblib')

movie_det.shape

# Transforming the DX data to a DataFrame and filling missing values

%%time
showings = pd.DataFrame()
for file in os.listdir(folder):
    temp = pd.read_json(folder+file, encoding='utf-8')
    if temp.shape == (0, 0): continue
    temp.set_index('arr_nr', inplace=True)
    temp.columns = columns
    temp['location'] = int(file.split('.')[0])
    temp['room'] = temp.loc[:, 'location'] * 100 + temp.loc[:, 'room']
    showings = showings.append(temp)
# Truncates the outlying coverage values to the boundary values
showings.loc[showings.coverage > 1, showings.coverage.name] = 1
showings.loc[showings.coverage < 0, showings.coverage.name] = 0
# Combining the strings for date and time to a datetime object
showings.loc[:, showings.date.name] = showings.date.str[:10]
showings.loc[:, showings.showtime.name] = showings.date.str[:] + ' ' + showings.showtime.str[:]
showings.drop(columns=['date'], inplace=True)
showings.showtime = pd.to_datetime(showings.showtime, format='%Y-%m-%d %H:%M:%S')
# Strips the titles of versions as demonstrated above in the notebook
showings['stripped_title'] = list(map(lambda title: rm_mf(title), showings.full_title.str.lower()))

## My most inefficient code ever

Running this would have taken a week

```python
for _, row in showings.iterrows():
    showings.loc[
        (index)
        & (showings.showtime == row['showtime']) 
        & (showings.sold == row['sold'])
        & (showings.std_movienr == row['std_movienr']),
        'full_title'
    ] = rm_mf(row['full_title'].lower())
```

While this code that does the exact same took 31.1 seconds

```python
showings.stripped_title = list(map(lambda title: rm_mf(title), showings.full_title.str.lower()))
```

## Remove NaN values

For some unknown reason we have two values for number of seats, these are genrally the same, but some are missing.
If one is missing then that value is set equal to the other, if both are missing the values are set to sold divided by coverage.

If the coverage is missing the missing values are set to sold divided by seats.

%%time
showings.loc[showings.seats.isna(), showings.seats.name] = showings.loc[showings.seats.isna(), showings.seats_b.name]
showings.loc[(showings.seats.isna()) & (showings.coverage.notna()), showings.seats.name] = (
    showings.loc[(showings.seats.isna()) & (showings.coverage.notna()), showings.sold.name] 
    / showings.loc[(showings.seats.isna()) & (showings.coverage.notna()), showings.coverage.name])
showings.loc[showings.seats_b.isna(), showings.seats_b.name] = showings.loc[showings.seats_b.isna(), showings.seats.name]
showings.loc[(showings.seats == 0) & (showings.seats_b != 0), showings.seats.name]   = showings.loc[(showings.seats == 0) & (showings.seats_b != 0), showings.seats_b.name]
showings.loc[(showings.seats != 0) & (showings.seats_b == 0), showings.seats_b.name] = showings.loc[(showings.seats != 0) & (showings.seats_b == 0), showings.seats.name]

# Set abnormally high numbers to approximately the median
showings.loc[showings.seats > 700, showings.seats.name] = np.random.randint(100, 300)
showings.loc[showings.seats_b > 700, showings.seats_b.name] = np.random.randint(100, 300)

while showings.coverage.isna().sum() > 0:
    if showings.loc[showings.coverage.isna(), 'seats'].sum() == 0 or showings.loc[showings.coverage.isna(), 'seats_b'].sum() == 0: break
    try:
        showings.loc[showings.coverage.isna(), 'coverage'] = showings.loc[showings.coverage.isna(), 'sold'] / showings.loc[showings.coverage.isna(), 'seats']
    except:
        showings.loc[showings.coverage.isna(), 'coverage'] = showings.loc[showings.coverage.isna(), 'sold'] / showings.loc[showings.coverage.isna(), 'seats_b']
        continue

showings.loc[showings.coverage > 1, showings.coverage.name] = 1
showings.loc[showings.coverage < 0, showings.coverage.name] = 0

showings.dropna(inplace=True)
showings.drop(showings.loc[(showings.seats == 0) & (showings.seats_b == 0)].index, inplace=True)

showings.isna().sum()

dump(showings, data_folder+'showings.joblib')

showings = load(data_folder+'showings.joblib')

showings.tail(10)
