==> notebooks/01_preprocess/02_data_aggregation.ipynb <==
# Notebook for aggregating data from different sources

### Sources:
- DX
- TMDb

import numpy as np
import pandas as pd
from joblib import load, dump
from tmdbv3api import TMDb, Movie

### Establishing a connection with the TMDb API

with open('C:/Users/marku/OneDrive/Skrivebord/TMDBkeys.txt') as keys:
    application_key = keys.readline().strip()
tmdb = TMDb()
tmdb.api_key = application_key
tmdb.language = 'en'
movie = Movie()

data = '../data/'

movie_det = load(data+'movie_details.joblib')

# Combine DX and TMDb

showings = load(data+'showings.joblib')

showtimes = list(showings.loc[:, 'showtime'].dt.strftime('%Y'))
titles = list(showings.loc[:, 'stripped_title'])

%%time
start = time()
append = list()
l = len(showtimes)
for i, year in enumerate(showtimes):
    append.append(movie_det.loc[(movie_det.show_year == year) & (movie_det.title == titles[i])].to_dict(orient='list'))
    if time()-start > 60:
        print(i, '\t/\t', l)
        start = time()
#     if i == 4: break
# Removes the nested list from inside the dict
for i in append:
    for j in i:
        if len(i[j]) == 0: i[j] = np.nan
        else: i[j] = i[j][0]

append = pd.DataFrame(append)
append.head()

dump(append, data+'append.joblib')

append = load(data+'append.joblib')

%%time
showings.reset_index(inplace=True)
showings['tmdb_title'] = append.tmdb_title
showings['tmdb_genre_ids'] = append.tmdb_genre_ids
showings['tmdb_original_language'] = append.tmdb_original_language
showings['tmdb_original_title'] = append.tmdb_original_title
showings['tmdb_overview'] = append.tmdb_overview
showings['tmdb_popularity'] = append.tmdb_popularity
showings['tmdb_release_date'] = append.tmdb_release_date
showings['tmdb_vote_average'] = append.tmdb_vote_average
showings['tmdb_vote_count'] = append.tmdb_vote_count
showings['tmdb_id'] = append.tmdb_id
showings['tmdb_runtime'] = append.tmdb_runtime
showings['tmdb_budget'] = append.tmdb_budget
showings['tmdb_production_companies'] = append.tmdb_production_companies
showings['tmdb_revenue'] = append.tmdb_revenue
showings['tmdb_tagline'] = append.tmdb_tagline
showings['tmdb_cast'] = append.tmdb_cast
showings['tmdb_crew'] = append.tmdb_crew
showings['tmdb_keywords'] = append.tmdb_keywords
showings['imdb_id'] = append.imdb_id

dump(showings, data+'showings.joblib')

showings = load(data+'showings.joblib')

showings.isna().sum()

showings.dropna(subset=['tmdb_title'], inplace=True)

showings.isna().sum()

dump(showings, data+'showings_dropped_NaN.joblib')
